// src/api/orval/model/userQuotaPlanDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface UserQuotaPlanDto {
  id?: number;
  userId?: number;
  quotaPlanId?: number;
  effectiveFrom?: string;
  /** @nullable */
  effectiveTo?: string | null;
  /** @nullable */
  assignedBy?: number | null;
  /** @nullable */
  note?: string | null;
}