// src/api/orval/model/createOrUpdateUserQuotaPlanRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface CreateOrUpdateUserQuotaPlanRequest {
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  id?: number;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  userId?: number;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  quotaPlanId?: number;
  effectiveFrom?: string;
  /** @nullable */
  effectiveTo?: string | null;
  /**
   * @minimum 0
   * @maximum 2147483647
   * @nullable
   */
  assignedBy?: number | null;
  /**
   * @maxLength 256
   * @nullable
   */
  note?: string | null;
}