// src/api/orval/model/getByTelegramIdMessagesResponseApiResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { GetByTelegramIdMessagesResponse } from "./getByTelegramIdMessagesResponse";

export interface GetByTelegramIdMessagesResponseApiResponse {
  success?: boolean;
  /** @nullable */
  message?: string | null;
  data?: GetByTelegramIdMessagesResponse;
}