// src/api/orval/model/messageDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export interface MessageDto {
  id?: number;
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  messageText?: string | null;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  fileId?: string | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  filePath?: string | null;
  receivedAt?: string;
  createDate?: string;
  lastUpdate?: string;
}